Add files via upload
This is a healthcare assistant. It is a 2 in 1 program. It has a Health Assistant, which has three options: Healthy, Sick and Very Sick. It gives food recommendations according to your health status. It will suggest some recipes too. It also has a BMI Calculator built in so that you can also calculate your BMI.